Lint, import updates

+++ frontend/src/editor/IntroPrompt.tsx
@@ -258,7 +258,7 @@ export function IntroPrompt() {
         return;
       }
 
-      const loadRes = saveResult.value as Uint8Array;
+      const loadRes = saveResult.value;
 
       if (!loadRes || loadRes.byteLength === 0) {
         console.error("Download failed: Generated map data is empty");
